Added systemcontroller

+++ controllers/group_controller.js
@@ -12,8 +12,8 @@ module.exports = {
 				.contentType('application/json')
 				.send(groups);
 		}).catch((err) => {
-			console.log('err on finding groups');
-			console.log(err);
+			console.error('err on finding groups');
+			console.error(err);
 			res.sendStatus(500);
 		})
 	},
@@ -28,29 +28,30 @@ module.exports = {
 					res.sendStatus(200);
 				}).catch((err) => {
 					res.sendStatus(500);
-					console.log('err on adding group');
-					console.log(err);
+					console.error('err on adding group');
+					console.error(err);
 				})
 			}
 		}).catch((err) => {
 			res.sendStatus(500);
-			console.log('err on finding group');
-			console.log(err);
+			console.error('err on finding group');
+			console.error(err);
 		})
 	},
 	
 	async addDeviceToGroup(device) {
 		let group = await Group.findOne({_id: device.groupId})
 			.catch(err => {
-				console.log(err);
+				console.error('err on finding group');
+				console.error(err);
 			});
 		
 		if (group.devices.length === 0 || !group.devices.includes(device._id)) {
 			Group.updateOne({_id: device.groupId}, {$push: {devices: device._id}}).then(() => {
 				return true;
 			}).catch((err) => {
-				console.log('err on adding to group');
-				console.log(err);
+				console.error('err on adding to group');
+				console.error(err);
 				return false;
 			})
 		} else {
@@ -64,21 +65,21 @@ module.exports = {
 		
 		let group = await Group.findOne({_id: _id});
 		
-		console.log(group);
-		
-		if (group.devices.length > 0) {
-			group.devices.forEach(device => {
-				setGroupToDefault(device);
-			})
+		if(group !== null) {
+			if (group.devices.length > 0) {
+				group.devices.forEach(device => {
+					setGroupToDefault(device);
+				});
+			}
+			
+			Group.findOneAndDelete({_id: _id}).then(() => {
+				res.sendStatus(200);
+			}).catch((err) => {
+				res.sendStatus(500);
+				console.error('err on deleting group');
+				console.error(err);
+			});
 		}
-		
-		Group.findOneAndDelete({_id: _id}).then(() => {
-			res.sendStatus(200);
-		}).catch((err) => {
-			res.sendStatus(500);
-			console.log('err on deleting group');
-			console.log(err);
-		})
 	},
 	
 	async removeDeviceFromGroup(device) {
@@ -88,8 +89,8 @@ module.exports = {
 			Group.updateOne({_id: device.groupId}, {$pull: {devices: device._id}}).then(() => {
 				return true;
 			}).catch((err) => {
-				console.log('err on deleting device from group');
-				console.log(err);
+				console.error('err on deleting device from group');
+				console.error(err);
 				return false;
 			});
 		} else {
@@ -104,8 +105,8 @@ module.exports = {
 		Device.findOneAndUpdate({_id: _id}, {groupId: 0}).then(() => {
 			return true;
 		}).catch((err) => {
-			console.log('err on setting group to default');
-			console.log(err);
+			console.error('err on setting group to default');
+			console.error(err);
 			return false;
 		});
